tested delete and put

+++ application.py
@@ -40,12 +40,16 @@ def add_drink():
     db.session.add(drink)
     db.session.commit()
     return {'id': drink.id}
-@app.route('/drinks/<id>', methods= ['DELETE'])
+@app.route('/drinks/<id>', methods=['DELETE'])
 def del_drink(id):
-    drink= Drink.query.get_or_404(id)
+    drink = Drink.query.get(id)
+    if drink is None:
+        return {"error": "Drink not found"}, 404
+
     db.session.delete(drink)
     db.session.commit()
-    return {"Message": "deletion done"}
+    return {"message": "Deletion successful"}
+
 @app.route('/drinks/<id>', methods= ['PUT'])
 def up_drink(id):
     drink= Drink.query.get_or_404(id)
